image_generator/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
openai>=1.0.0
Pillow>=10.0.0
redis>=5.0.0
python-dotenv>=1.0.0
